Wrap function to node js script
